Add MovieGenreController and update layout views
This commit adds a new MovieGenreController with index, create, store, show, edit, update and destroy methods to display, create, store, update and destroy MovieGenre. Additionally, the layout views have been updated to wrap the @yield content with a 'd-flex justify-content-between' class to create space at the right-hand side of the content and remove some html comments and sections that were not needed.